refactor: TaskCategoryBadge to domain types folder
